Prepare for releasing

+++ modules/circe/src/main/scala/circe/CirceSolver.scala
@@ -1,11 +1,27 @@
+/*
+ * Copyright 2023 Filippo De Luca
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package com.filippodeluca.jsonpath
 package circe
 
-import cats.syntax.all._
+import cats.syntax.all.*
 
 import io.circe.Json
 
-import com.filippodeluca.jsonpath.ast._
+import com.filippodeluca.jsonpath.ast.*
 
 object CirceSolver {
 
@@ -85,7 +101,7 @@ object CirceSolver {
             .flatMap { index =>
               target.asArray
                 .flatMap { arr =>
-                  arr.get(index)
+                  arr.get(index.toLong)
                 }
             }
 
@@ -118,7 +134,7 @@ object CirceSolver {
             )
 
             Json.fromValues(range.toVector.mapFilter { idx =>
-              val value = array.get(idx)
+              val value = array.get(idx.toLong)
 
               value
             })
